+++ examples/java-jpms/closed-consumer/src/main/java/module-info.java
@@ -1,0 +1,3 @@
+module io.github.frikit.krandom.examples.jpms.closed {
+    requires io.github.frikit.krandom;
+}