+++ examples/java-jpms/open-consumer/src/main/java/module-info.java
@@ -1,0 +1,5 @@
+module io.github.frikit.krandom.examples.jpms.open {
+    requires io.github.frikit.krandom;
+
+    opens io.github.frikit.krandom.examples.jpms.openconsumer to io.github.frikit.krandom;
+}